main renderer - move treasury tasks to context
