+++ packages/renderer/src/contexts/main/TreasuryApi/defaults.ts
@@ -1,0 +1,9 @@
+// Copyright 2025 @polkadot-live/polkadot-live-app authors & contributors
+// SPDX-License-Identifier: GPL-3.0-only
+/* eslint-disable @typescript-eslint/no-unused-vars, @typescript-eslint/no-empty-function */
+
+import type { TreasuryApiContextInterface } from './types';
+
+export const defaultTreasuryApiContext: TreasuryApiContextInterface = {
+  handleInitTreasury: () => new Promise(() => {}),
+};
