+++ packages/renderer/src/contexts/main/TreasuryApi/types.ts
@@ -1,0 +1,6 @@
+// Copyright 2025 @polkadot-live/polkadot-live-app authors & contributors
+// SPDX-License-Identifier: GPL-3.0-only
+
+export interface TreasuryApiContextInterface {
+  handleInitTreasury: (ev: MessageEvent) => Promise<void>;
+}
