+++ packages/renderer/src/contexts/main/index.ts
@@ -14,4 +14,5 @@ export * from './IntervalTasksManager';
 export * from './LedgerSigner';
 export * from './Manage';
 export * from './Subscriptions';
+export * from './TreasuryApi';
 export * from './WalletConnect';
