+++ packages/renderer/src/contexts/openGov/Treasury/index.tsx
@@ -2,9 +2,9 @@
 // SPDX-License-Identifier: GPL-3.0-only
 
 import * as defaults from './default';
+import BigNumber from 'bignumber.js';
 import { ConfigOpenGov, formatBlocksToTime } from '@polkadot-live/core';
 import { createContext, useContext, useRef, useState } from 'react';
-import BigNumber from 'bignumber.js';
 import { rmCommas } from '@w3ux/utils';
 import { chainCurrency } from '@polkadot-live/consts/chains';
 import type { AnyData } from '@polkadot-live/types/misc';
